+++ tests/Pinta.Effects.Tests/Mocks/MockPalette.cs
@@ -1,4 +1,5 @@
 using System;
+using System.Collections.ObjectModel;
 using Cairo;
 using Pinta.Core;
 
@@ -8,6 +9,12 @@ internal sealed class MockPalette : IPaletteService
 {
 	public Color PrimaryColor { get; set; } = new (0, 0, 0); // Black
 	public Color SecondaryColor { get; set; } = new (1, 1, 1); // White
+
+	public Palette CurrentPalette
+		=> throw new NotImplementedException ();
+
+	public ReadOnlyCollection<Color> RecentlyUsedColors
+		=> throw new NotImplementedException ();
 
 #pragma warning disable CS0067 // The event 'MockPalette.PrimaryColorChanged' is never used
 	public event EventHandler? PrimaryColorChanged;